backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
feedparser==6.0.11
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.32.3
pytz==2024.1
telethon==1.36.0
beautifulsoup4==4.12.3

